refactor: move HomePage to new directory structure and update import path
